feat: added TS to index and router
