+++ server/router.ts
@@ -1,0 +1,17 @@
+import {Router} from 'express';
+const reports = require('./controllers/reports.controller');
+const authMiddleware = require('./middlewares/auth');
+
+const router = Router();
+//Reports routes
+router.get('/allreports', reports.allReports);
+
+router.get('/getreport/:id', reports.getReport);
+
+router.post('/postreport', reports.newReport);
+
+router.patch('/editreport', authMiddleware, reports.editReport);
+
+router.delete('/deletereport/:id', authMiddleware, reports.deleteReport);
+
+export default router;
